SettingsObserver implemented
Because settings can be changed in multiple places, I decided to create a SettingsObserver class/object that maintains a copy of the currentSettings. Every time it is called, it checks for changes in the currentSettings vs its _settings and as needed invokes the applySettings method of the relevant controller.

Unfortunately, the way the knobs are programmed, their position will immediately undo any changes made from BT or USB. Need knobs to know if they've been turned and only change settings in that event.